Tests for structure value types.
